# refactor(spanner): move Base64 codecs to common