--- a/google/cloud/google_cloud_cpp_common_unit_tests.bzl
+++ b/google/cloud/google_cloud_cpp_common_unit_tests.bzl
@@ -26,6 +26,7 @@ google_cloud_cpp_common_unit_tests = [
     "internal/algorithm_test.cc",
     "internal/api_client_header_test.cc",
     "internal/backoff_policy_test.cc",
+    "internal/base64_transforms_test.cc",
     "internal/big_endian_test.cc",
     "internal/compiler_info_test.cc",
     "internal/credentials_impl_test.cc",