… google/cloud/google_cloud_cpp_common.bzl
… google/cloud/spanner/bytes_test.cc