I think these are better than the OpenSSL-based ones in the storage
library. In this step I refactored the core functions to the common
libraries.

I changed the encoder class to make it easier / safer to use outside the
`spanner::Bytes` class.

I moved most of the tests too, I did not think it made sense to
duplicate the tests.

Part of the work for #6946

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/googleapis/google-cloud-cpp/6954)
<!-- Reviewable:end -->