@coryan 2021-07-08
I do not expect any significant changes in performance from these changes, but for completeness I ran the benchmark.

I compiled on my machine with GCC 10 and `-O3` I got:

## Before

```
cmake-build-release/google/cloud/spanner/bytes_benchmark --benchmark_min_time=30
2021-07-08T14:15:52+00:00
Running cmake-build-release/google/cloud/spanner/bytes_benchmark
Run on (96 X 2000.16 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x48)
  L1 Instruction 32 KiB (x48)
  L2 Unified 1024 KiB (x48)
  L3 Unified 39424 KiB (x2)
Load Average: 3.30, 5.29, 6.68
-----------------------------------------------------------------------
Benchmark             Time             CPU   Iterations UserCounters...
-----------------------------------------------------------------------
BM_BytesCtor       6313 ns         6312 ns      6573792 bytes_per_second=236.304M/s
BM_BytesGet        4663 ns         4662 ns      9178377 bytes_per_second=427.116M/s
```

## After

```
cmake-build-release/google/cloud/spanner/bytes_benchmark --benchmark_min_time=30
2021-07-08T14:21:33+00:00
Running cmake-build-release/google/cloud/spanner/bytes_benchmark
Run on (96 X 2000.16 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x48)
  L1 Instruction 32 KiB (x48)
  L2 Unified 1024 KiB (x48)
  L3 Unified 39424 KiB (x2)
Load Average: 9.98, 6.66, 6.54
-----------------------------------------------------------------------
Benchmark             Time             CPU   Iterations UserCounters...
-----------------------------------------------------------------------
BM_BytesCtor       6149 ns         6148 ns      6895907 bytes_per_second=242.606M/s
BM_BytesGet        4639 ns         4638 ns      8922267 bytes_per_second=429.32M/s
```